cardno change  from double to long

+++ Code/FraudDetection/fraudanalytics-dto/src/main/java/com/masteklabs/fraudanalytics/common/dto/PredictionResult.java
@@ -62,7 +62,7 @@ public class PredictionResult implements Serializable{
 	public Double getCardNumber() {
 		return cardNumber;
 	}
-	public void setCardNumber(Double cardNumber) {
+	public void setCardNumber(Long cardNumber) {
 		this.cardNumber = cardNumber;
 	}
